Added a German translation for the KiCad Windows Installer (#40)
* Add German translation to KiCad Windows Installer

Adding a first try of a German localization for the Windows Installer
for KiCad. The look and feel isn't tested as I haven't built any Windows
installer so probably some adjustments will be needed.

* Adding German to the entries of the Installer

While adding this keep up a alphabetical order within the file like done
on the KiCad website for better readability.