Remove libs and docs
This is so we can have lite builds for the nightlies... This is just a
hacky way to implement it for now. We still need to decide what we
really need. It would be good to be able to make a normal and a lite
build with the same pkg.tar.xz.