Add a fallback user directory
For the first time I see the issue some people have been reported about
the user not showing up and the build fails.

It seems that depending on what type of windows user you have the setup
is initilized slightly different in msys2, so I will solve this with a
fallback user name of "user". The user name is not really important,
something just need to be defined.